backend/requirements.txt
asgiref==3.8.1
Django==5.0.12
django-cors-headers==4.7.0
django-extensions==3.2.3
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mssql-django==1.5
PyJWT==2.9.0
pyodbc==5.2.0
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
